Feature: fetch ASN for all nodes
Problem: we want to measure the decentralization of the network,
i.e. identify where the nodes are hosted.

Solution: we now retrieve the Autonomous System Number (ASN) of
each node. This ID allows to determine where the servers are hosted
(if they are not proxied).

To do so, we use pyasn. pyasn is a Python module that can look up
ASNs from a local ASN database, meaning that it does not require
to access a third-party service once the DB is downloaded.

Currently, we refresh the database at startup and then once every
24 hours.