Feature: use ASN DB file mtime to determine refresh
Problem: the current refresh mechanism uses a variable to determine
the last update time of the ASN DB file. On consecutive application
runs, a DB refresh (=download + conversion) is always required.

Solution: use the modification time of the ASN DB file instead.
The DB is not refreshed anymore if the DB file exists and is not
older than the specified period (1 day by default).