Fix: Event loop is closed
Problem: the scoring system sometimes hangs when posting to Aleph.
We suspect that the issue is caused by the reuse of the default
aiohttp session of `aleph-client`. The scheduling system creates
new asyncio event loops but the client session remains configured
with the first one.

Solution: create a new aiohttp session object whenever
pushing scores to Aleph instead of reusing the default one.